refactor(socket): split Response into smaller files
