--- zio-http/src/main/scala/zhttp/socket/SocketError.scala
@@ -1,8 +1,0 @@
-package zhttp.socket
-
-sealed trait SocketError extends Throwable
-
-object SocketError {
-  case object UnknownMessage extends SocketError
-  def unknown: SocketError = UnknownMessage
-}
